implement

+++ Source/datalayer/TradeInfo.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface TradeInfo {
     String bookNumber=null;
-    int priceSum=0;
+    Double priceSum=0.00;
     String Date=null;
     int Count=0;
 }
